+++ java/pom.xml
@@ -4,11 +4,11 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>io.github.ros</groupId>
-  <artifactId>rostools</artifactId>
+  <groupId>net.danielgill</groupId>
+  <artifactId>railostools</artifactId>
   <version>1.0.0</version>
 
-  <name>rostools</name>
+  <name>railostools</name>
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
