+++ java/src/main/java/net/danielgill/railostools/Metadata.java
@@ -1,4 +1,4 @@
-package io.github.ros;
+package net.danielgill.railostools;
 
 import java.util.List;
 import java.util.Map;
