+++ java/src/main/java/net/danielgill/railostools/MetadataBuilder.java
@@ -1,4 +1,4 @@
-package io.github.ros;
+package net.danielgill.railostools;
 
 import java.io.File;
 import java.util.Map;
